requirements.txt
aiogram==3.10.0
aiosqlite==0.19.0
python-dotenv==1.0.0
gspread==6.0.0